box2 (beam_graphics v0.1.0)

2D Axis-aligned Bounding Box (AABB).

To be written.

Functions

contains/2

-spec contains(graphics:box2(), graphics:vector2()) -> boolean().

Checks if a point is inside a box.

It returns true if the point is inside the box, false otherwise.

from_points/1

-spec from_points([graphics:vector2()]) -> graphics:box2().

Creates a 2D box from a list of points.

It computes the minimum and maximum points of the list of points and returns the bounding box that contains all the points.

from_vertices(Vertices)

-spec from_vertices([graphics:vertex2()]) -> graphics:box2().

Creates a 2D box from a list of vertices.

Same from from_points/1 but it extracts the position of the vertices.

intersects/2

-spec intersects(graphics:box2(), graphics:box2()) -> boolean().

Checks if two boxes intersect.

It returns true if the two boxes intersect, false otherwise.