color (beam_graphics v0.1.0)

The color API.

It provides functions to create and manipulate colors. Each color is represented as a tuple of four integers, corresponding to the red, green, blue, and alpha channels. It defines the RGBA color type, where each channel ranges from 0 to 255.

Types

channel()

-type channel() :: float().

Functions

alpha(Color)

-spec alpha(graphics:color()) -> integer().

The alpha component of a color.

It returns the alpha component of the color, which is an integer ranging from 0 to 255.

argb(Alpha, Red, Green, Blue)

-spec argb(channel(), channel(), channel(), channel()) -> graphics:color().

Create a color from ARGB components.

It takes four integers representing the alpha, red, green and blue components of the color, each ranging from 0 to 255, and returns a color tuple.

blue(Color)

-spec blue(graphics:color()) -> integer().

The blue component of a color.

It returns the blue component of the color, which is an integer ranging from 0 to 255.

green(Color)

-spec green(graphics:color()) -> integer().

The green component of a color.

It returns the green component of the color, which is an integer ranging from 0 to 255.

red(Color)

-spec red(graphics:color()) -> integer().

The red component of a color.

It returns the red component of the color, which is an integer ranging from 0 to 255.

rgb(Red, Green, Blue)

-spec rgb(channel(), channel(), channel()) -> graphics:color().

Create a color from RGB components.

It creates a color from three integers representing the red, green and blue components, each ranging from 0 to 255, and returns a color tuple with alpha set to 255.

rgba(Red, Green, Blue, Alpha)

-spec rgba(channel(), channel(), channel(), channel()) -> graphics:color().

Create a color from RGBA components.

It takes four integers representing the red, green, blue and alpha components of the color, each ranging from 0 to 255, and returns a color tuple.

to_integer/1

-spec to_integer(bml:color()) -> integer().

Convert a color to an integer.

It converts a color to a 32-bit integer in 16xRRGGBBAA format.