graphics (beam_graphics v0.1.0)

To be written.

To be written.

Types

angle()

-type angle() :: float().

An angle in radians.

To be written.

box2()

-type box2() :: {Min :: vector2(), Max :: vector2()}.

To be written.

To be written.

box3()

-type box3() :: {Min :: vector3(), Max :: vector3()}.

To be written.

To be written.

camera2()

-type camera2() :: term().

To be written.

To be written.

camera3()

-type camera3() :: term().

To be written.

To be written.

color()

-type color() ::
          {Red :: color:channel(),
           Green :: color:channel(),
           Blue :: color:channel(),
           Alpha :: color:channel()}.

A RGBA color.

To be written.

matrix3()

-type matrix3() ::
          {M11 :: float(),
           M21 :: float(),
           M31 :: float(),
           M12 :: float(),
           M22 :: float(),
           M32 :: float(),
           M13 :: float(),
           M23 :: float(),
           M33 :: float()}.

A 3x3 matrix.

A 3x3 grid of numbers typically used to represent 2D transformations in the Euclidean plane.

column majoor order xxx

matrix4()

-type matrix4() ::
          {M11 :: float(),
           M21 :: float(),
           M31 :: float(),
           M41 :: float(),
           M12 :: float(),
           M22 :: float(),
           M32 :: float(),
           M42 :: float(),
           M13 :: float(),
           M23 :: float(),
           M33 :: float(),
           M43 :: float(),
           M14 :: float(),
           M24 :: float(),
           M34 :: float(),
           M44 :: float()}.

A 4x4 matrix.

A 4x4 grid of numbers typically used to represent 3D transformations in the Euclidean space.

column majoor order xxx

mesh2()

-type mesh2() :: mesh2:object().

2D mesh object.

A collection of 2D vertices that can be rendered on a surface.

Use the mesh2 module to create and manipulate 2D meshes.

mesh3()

-type mesh3() :: mesh3:object().

3D mesh object.

A collection of 3D vertices that can be rendered on a surface.

Use the mesh3 module to create and manipulate 3D meshes.

primitive_type()

-type primitive_type() ::
          points | lines | line_strip | line_loop | triangles | triangle_strip | triangle_fan.

To be written.

To be written.

program()

-type program() :: program:object().

To be written.

To be written.

shape2()

-type shape2() ::
          #shape2{meshes :: [{mesh2:mesh(), graphics:primitive_type(), graphics:vertex_count()}],
                  matrix :: graphics:matrix3(),
                  texture :: no_texture | graphics:texture()}.

To be written.

To be written.

shape3()

-type shape3() ::
          #shape3{meshes :: [{mesh3:mesh(), graphics:primitive_type(), graphics:vertex_count()}],
                  matrix :: graphics:matrix4(),
                  texture :: no_texture | graphics:texture()}.

To be written.

To be written.

surface()

-type surface() :: surface:object().

To be written.

To be written.

texture()

-type texture() :: texture:object().

To be written.

To be written.

vector2()

-type vector2() :: {X :: float(), Y :: float()}.

A 2D vector.

A pair of numbers typically used to represent 2D positions and directions in the Euclidean plane.

Note that it defines operations with 3x3 matrices by assuming an invisible third component set to 0.0 (also called the homogeneous coordinate).

vector3()

-type vector3() :: {X :: float(), Y :: float(), Z :: float()}.

A 3D vector.

A triplet of numbers typically used to represent 3D positions and directions in the Euclidean space.

Note that it defines operations with 4x4 matrices by assuming an invisible fourth component set to 0.0 (also called the homogeneous coordinate).

vertex2()

-type vertex2() :: {Position :: vector2(), Color :: color(), U :: float(), V :: float()}.

To be written.

To be written.

vertex3()

-type vertex3() :: {Position :: vector3(), Color :: color(), U :: float(), V :: float()}.

To be written.

To be written.

vertex_count()

-type vertex_count() :: non_neg_integer().

To be written.

To be written.

Functions

initialize(Display)

-spec initialize(egl:display()) -> ok.

To be written.

To be written.

terminate()

-spec terminate() -> ok.

To be written.

To be written.