graphics_context (beam_graphics v0.1.0)

The root graphics context.

It's a background process that holds an OpenGL context which is shared with the OpenGL context of all graphics:surface/0 objects. It's responsible for creating, manipulating, and destroying OpenGL resources such as buffers and textures which are shareable resources (and therefore can be used by all surfaces).

The OpenGL context is always active within the process. The sole responsibility of the process is to handle OpenGL resources and operations, and frees up when process owning them die.

The root graphics context serves as the foundation of the graphics library and therefore must be started before any other graphics-related operations can take place.

ok = graphics_context:start().

Note that it's done by the graphics:initialize/0 function (which should be used instead).

To access the inner OpenGL context, use the graphics_context:inner_context/0 function.

{ok, Context} = graphics_context:inner_context(),

Unless you intend to use your own OpenGL calls, you do not need to retrieve the inner OpenGL context. (rephrase).

Later...

ok = graphics_context:stop().

It's indirectly performed by the graphics:terminate/0 function.

OpenGL resources and ownership

The sole responsibility of the process is to holds shared OpenGL resources which implies a dependency of those resources on the root graphics context.

About "mesh" operations

To be written.

About "texture" operations

The texture-related functions allow to create, manipulate, and destroy OpenGL textures (on the root OpenGL context). Together, those functions allow the implementation of the graphics:texture/0 object.

texture_new/3 - Create an OpenGL texture (and initialize it, first allocation).

texture_set_data/3 - Set the data of an OpenGL texture (re-allocation).

texture_data/3 - Read the data (or a subset) of an OpenGL texture.

texture_update_data/3 - Update the data of an OpenGL texture (no re-allocation).

texture_destroy/1 - Destroy an OpenGL texture.

Notice how they reflect the semantics of the OpenGL API (uninitialized/allocation/re-allocation) with one constraint: the texture is always initialized (it must be at least 1x1 pixel in size).

{ok, [Texture]} = gl:gen_textures(1),
ok = gl:bind_texture(texture_2d, Texture),
ok = gl:tex_image_2d()

Each pixel must be encoded in the following format:

R:8/unsigned byte
G:8/unsigned byte
B:8/unsigned byte
A:8/unsigned byte

Blabla.

About "shader program" operations

To be written.

XXX: Can be optimized in various way. For instance, if keeping the a release 
     function per resource is too expensive, it can implement a "register 
    "resource type" mechanism.
XXX: The start/x function should return infos about the OpenGL context, such as
     the version, the vendor, etc.
XXXX: verify if a pbuffer surface is needed (or if passing no_surface is enough)

XXX: Consider adding "kill" option to stop/x function in order to kill owners of
     resources before freeing the resources.

Types

resource_acquire_fun()

-type resource_acquire_fun() :: fun(() -> {ok, resource_id(), resource_release_fun()} | {error, term()}).

To be written.

To be written.

resource_id()

-type resource_id() :: term().

resource_release_fun()

-type resource_release_fun() :: fun((resource_id()) -> ok).

To be written.

To be written.

Functions

acquire_resource(AcquireFun)

-spec acquire_resource(resource_acquire_fun()) -> {ok, resource_id()} | {error, term()}.

To be written.

To be written.

execute_commands(Commands)

-spec execute_commands(fun(() -> term())) -> term().

To be written.

To be written.

handle_message/2

handle_request/3

initialize/1

inner_context()

-spec inner_context() -> egl:context().

The inner OpenGL context.

It returns the inner OpenGL context which is shared by the OpenGL context of all surfaces.

inner_resources()

-spec inner_resources() -> #{resource_id() := pid()}.

To be written.

To be written.

release_resource(ResourceId)

-spec release_resource(resource_id()) -> ok | invalid_resource_id.

To be written.

To be written.

start(Display)

-spec start(egl:display()) -> worker:start_ret().

To be written.

To be written.

stop()

-spec stop() -> ok.

To be written.

To be written.

terminate/2

transfer_ownership(ResourceId, Owner)

-spec transfer_ownership(resource_id(), pid()) -> ok | invalid_resource_id.

To be written.

To be written.