matrix4 (beam_graphics v0.1.0)

4x4 Matrix

A 4x4 matrix is a grid of numbers that is typically used to represent 3D transformations in the Euclidean space.

It can also be used as a 2x3 matrix (by ignoring the third row and assuming its value is [0, 0, 1], also called the homogeneous ' coordinates) and therefore define operations with 2D vectors.

It's an unintuitive mathematical structure that is hard to manipulate directly. Instead, use the transform2 and view2 modules that provides higher-level operation to define your 2D transformations. They both wrap a a 4x4 matrix. of that, it's recommended to use the transform2 module that provides a more convenient interface to manipulate 2D transformations. The data structure of a 4x4 matrix simply is a tuple of 9 floats where the elements represent the 2D array from top to bottom, left to right. Therefore, 4x4 matrices can naturally be created with the tuple syntax.

M = {
    1.0, 2.0, 3.0,
    4.0, 5.0, 6.0,
    7.0, 8.0, 9.0
}.

To access the elements of a 4x4 matrix, use either the element/2 or element3 function.

1.0 = matrix3:elem(M, 1).
4.0 = matrix3:elem(M, 4).
1.0 = matrix3:element(M, 1, 1).
3.0 = matrix3:element(M, 3, 1).
7.0 = matrix3:element(M, 1, 3).

A 4x4 matrix where all the elements are set to zero is called a zero matrix, which can be conveniently created with the zero/0 function. Also a zero 4x4 matrix with a 1.0 in the bottom right corner is called an identity matrix, which can be conveniently created with the identity/0 function.

{
    0.0, 0.0, 0.0,
    0.0, 0.0, 0.0,
    0.0, 0.0, 0.0
} = matrix3:zero().
{
    0.0, 0.0, 0.0,
    0.0, 0.0, 0.0,
    0.0, 0.0, 0.0
} = matrix3:identity().

Constants are also defined to represent the zero and identity 4x4 matrix. You must include the beam_graphics.hrl header to use them.

{
    0.0, 0.0, 0.0,
    0.0, 0.0, 0.0,
    0.0, 0.0, 0.0
} = ?MATRIX_4x4_ZERO.
{
    0.0, 0.0, 0.0,
    0.0, 0.0, 0.0,
    0.0, 0.0, 0.0
} = ?MATRIX_4x4_IDENTITY.

The common mathematical operations are also implemented.

Types

vector4()

-type vector4() :: {float(), float(), float(), float()}.

Functions

add(Matrix1, Matrix2)

-spec add(graphics:matrix4(), graphics:matrix4()) -> graphics:matrix4().

To be written.

To be written.

column(Matrix, Column)

-spec column(graphics:matrix4(), Index :: 1..4) -> graphics:vector4().

To be written.

To be written.

columns(Matrix)

-spec columns(graphics:matrix4()) ->
                 {graphics:vector4(), graphics:vector4(), graphics:vector4(), graphics:vector4()}.

To be written.

To be written.

determinant(Matrix)

-spec determinant(graphics:matrix4()) -> float().

To be written.

To be written.

divide(Matrix, Divider)

-spec divide(graphics:matrix4(), float()) -> graphics:matrix4().

To be written.

To be written.

element(Matrix, Row, Column)

-spec element(graphics:matrix4(), integer(), integer()) -> float().

To be written.

To be written.

from_columns(Column1, Column2, Column3, Column4)

-spec from_columns(vector4(), vector4(), vector4(), vector4()) -> graphics:matrix4().

To be written.

To be written.

from_rows(Row1, Row2, Row3, Row4)

-spec from_rows(vector4(), vector4(), vector4(), vector4()) -> graphics:matrix4().

To be written.

To be written.

identity()

-spec identity() -> graphics:matrix4().

The identity 4x4 matrix.

It constructs a zero 4x4 matrix.

{1.0, 0.0, 0.0, 0.0, 0.0, 1.0, 0.0, 0.0, 0.0, 0.0, 1.0, 0.0, 0.0, 0.0, 0.0, 1.0} =
    matrix3:identity().

Note that the ?MATRIX4_IDENTITY macro can be used instead.

inverse(Matrix)

-spec inverse(graphics:matrix4()) -> graphics:matrix4() | undefined.

To be written.

To be written.

is_equal_to(Matrix1, Matrix2)

-spec is_equal_to(graphics:matrix4(), graphics:matrix4()) -> boolean().

To be written.

To be written.

is_equal_to(Matrix1, Matrix2, Epsilon)

-spec is_equal_to(graphics:matrix4(), graphics:matrix4(), float()) -> boolean().

To be written.

To be written.

is_identity(Matrix)

-spec is_identity(graphics:matrix4()) -> boolean().

To be written.

To be written.

is_orthogonal(Matrix)

-spec is_orthogonal(graphics:matrix4()) -> boolean().

To be written.

To be written.

is_symmetric(Matrix)

-spec is_symmetric(graphics:matrix4()) -> boolean().

To be written.

To be written.

is_zero(Matrix)

-spec is_zero(graphics:matrix4()) -> boolean().

To be written.

To be written.

lerp(Matrix1, Matrix2, T)

-spec lerp(graphics:matrix4(), graphics:matrix4(), float()) -> graphics:matrix4().

To be written.

To be written.

multiply(Matrix1, Matrix2)

-spec multiply(graphics:matrix4(), graphics:matrix4()) -> graphics:matrix4().

To be written.

To be written.

multiply_vector(Matrix, Vector)

-spec multiply_vector(graphics:matrix4(), graphics:vector3()) -> graphics:matrix4().

To be written.

To be written.

row(Matrix, Row)

-spec row(graphics:matrix4(), Index :: 1..4) -> graphics:vector4().

To be written.

To be written.

rows(Matrix)

-spec rows(graphics:matrix4()) ->
              {graphics:vector4(), graphics:vector4(), graphics:vector4(), graphics:vector4()}.

To be written.

To be written.

scale(Matrix, Scalar)

-spec scale(graphics:matrix4(), float()) -> graphics:matrix4().

To be written.

To be written.

smooth_lerp(Matrix1, Matrix2, T)

-spec smooth_lerp(graphics:matrix4(), graphics:matrix4(), float()) -> graphics:matrix4().

To be written.

To be written.

subtract(Matrix1, Matrix2)

-spec subtract(graphics:matrix4(), graphics:matrix4()) -> graphics:matrix4().

To be written.

To be written.

to_matrix3(Matrix)

-spec to_matrix3(graphics:matrix4()) -> graphics:matrix3().

To be written.

To be written.

transpose(Matrix)

-spec transpose(graphics:matrix4()) -> graphics:matrix4().

To be written.

To be written.

zero()

-spec zero() -> graphics:matrix4().

The zero 4x4 matrix.

It constructs a zero 4x4 matrix.

{0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0} =
    matrix4:zero().

Note that the ?MATRIX4_ZERO macro can be used instead.