mesh2 (beam_graphics v0.1.0)

2D Mesh

A 2D mesh is a collection of 2D vertices which are used for rendering. Each 2D vertex specifies a position in 2D space, a color and texture coordinates. They are rendered on a surface by using one of the following rendering primitives, and an optional texture.

For instance, the following will render a colored triangle.

Mesh = mesh:with_vertices([
    {100.0, 100.0, ?COLOR_RED, 0.0, 0.0},
    {700.0, 100.00, ?COLOR_GREEN, 0.0, 0.0},
    {700.0, 500.00, ?COLOR_BLUE, 0.0, 0.0}
]).
surface:render(Surface, Mesh, triangle_fan, no_texture).

How vertices are used to produce a visual output depends on the rendering primitive which is described in the OpenGL documentation.

Together with 3D meshes, meshes are the only mean for rendering. Anything else (.e.g., sprites, text, shapes, etc.) is a higher-level abstraction that indirectly uses meshes for rendering.

Mesh vertices must live in the GPU memory in order to be used for rendering, and therefore, reading and updating vertices have an associated cost. Avoid those operations as much as possible, and if frequent reading is required, pixels can be cached locally with the keep_copy option. It explains the local/remote semantics in the API. For instance, local_vertices/1 will read the vertices from a local copy (if it exists), while remote_image/1 will read the vertices from the GPU memory (which is more expensive).

While updating a 2D mesh should be avoided when not necessary, it's acceptable to frequently update its vertices. To help the GPU with handling the data in the most optimal way, you can specify a usage hint when setting the vertices.

To empty the mesh, you can call set_vertices(Mesh, []). If you need to "pre-allocate" the mesh. Fill it with "zero" vertices.

OpenGL Internals

It wraps a buffer object. To be written.

Use the handle/1 function to retrieve the OpenGL buffer ID. It's created by the root context, which you can retrieve with the graphics_context:inner_context/0 function.

XXX: The slice:offset in update_vertices and vertex functions is weird.
XXX: Clarify what happens if destroyed twice ?
XXX: Implement a merge/x or combine/x function to concatenate two meshes.
XXX: Update vertices functions are to be implemented.
XXX: Implement transfer "ownership".
XXX: Consider vector2() instead of X, Y.
XXX: Consider adding a sort of "normalized_float()" type for U V components.
XXX: Should it expose the other usage hint (OpenGL defines more than that).
XXX: Consider combining `no_copy | keep_copy` and usage_hint() into an "options".
XXX: Consider implementing:

-spec set_position(mesh3(), index(), vector3()) -> ok.
-spec set_color(mesh3(), index(), color_rgba()) -> ok.
-spec set_texcoord(mesh3(), index(), vector2()) -> ok.

XXX: About usage hints:

OpenGL Usage Hint	Meaning
GL_STATIC_DRAW (default)	Data set once, used many times (e.g., level geometry)
GL_DYNAMIC_DRAW	Data modified occasionally, used many times (e.g., animated meshes)
GL_STREAM_DRAW	Data modified every frame (e.g., particle systems)
GL_STATIC_READ	Data written by GPU (e.g., compute shader output)
GL_DYNAMIC_READ	Data occasionally written by GPU, read by CPU
GL_STREAM_READ	Data written by GPU every frame, read by CPU
GL_STATIC_COPY	Data written by GPU, used by GPU (rare)
GL_DYNAMIC_COPY	Data occasionally written by GPU, used by GPU
GL_STREAM_COPY	Data written by GPU every frame, used by GPU

Types

keep_copy()

-type keep_copy() :: no_copy | keep_copy.

To be written.

object()

-opaque object()

A 2D mesh object.

It wraps the OpenGL buffer ID and possibly a local copy of the vertices.

usage_hint()

-type usage_hint() :: static | dynamic | stream.

To be written.

vertex()

-type vertex() :: {X :: float(), Y :: float(), Color :: graphics:color(), U :: float(), V :: float()}.

To be written.

vertices()

-type vertices() :: [vertex()].

To be written.

Functions

destroy/1

-spec destroy(object()) -> ok.

To be written.

To be written.

gl_object/1

-spec gl_object(object()) -> gl:buffer().

To be written.

To be written.

local_vertex/2

-spec local_vertex(object(), slice:index()) -> undefined | out_of_range | {slice:offset(), vertex()}.

A vertex of the mesh.

It returns a specific vertex of the mesh, identified by a slice index.

Vertices = [
    {0.0, 0.0, ?COLOR_RED, 0.0, 0.0},
    {1.0, 0.0, ?COLOR_GREEN, 1.0, 0.0},
    {0.0, 0.0, ?COLOR_BLUE, 0.0, 0.0}
].
Mesh = mesh:with_vertices(Vertices, with_copy).

{0.0, 0.0, ?COLOR_RED, 0.0, 0.0} = mesh:local_vertices(Mesh, 0).
{0.0, 0.0, ?COLOR_BLUE, 0.0, 0.0} = mesh:local_vertices(Mesh, -1).

It returns undefined if no local copy is kept. If the index is out of range, it returns out_of_range.

local_vertices/1

-spec local_vertices(object()) -> undefined | vertices().

The locally cached vertices of the mesh.

It returns the local copy of the vertices that are currently set in the mesh, if a copy was with the setIf no local vertex data is cached, it returnsundefined. Uselocal_vertices/2to only retrieve a subset of the vertices using the slice notation. It returnsundefined` if no local copy is kept. By default, no local copy is kept

local_vertices/2

-spec local_vertices(object(), slice:range()) -> undefined | no_range | {slice:offset(), vertices()}.

A slice of the locally cached vertices of the mesh.

It returns the local copy of the vertices that are currently set in the mesh, if a copy was with the setIt returnsundefined` if no local copy is kept. By default, no local copy is kept

remote_vertex/2

-spec remote_vertex(object(), slice:index()) -> out_of_range | {slice:offset(), vertex()}.

To be written.

To be written.

remote_vertices/1

-spec remote_vertices(object()) -> vertices().

To be written.

To be written.

remote_vertices/2

-spec remote_vertices(object(), slice:range()) -> no_range | {slice:offset(), vertices()}.

To be written.

To be written.

set_vertices(Mesh, Vertices)

-spec set_vertices(object(), vertices()) -> {ok, object()} | out_of_memory.

Set the vertices of the mesh.

It sets the vertices foobar.

It's equivalent to set_vertices(Mesh, Vertices, static).

See set_vertices/3 for more details.

set_vertices(Mesh, Vertices, UsageHint)

-spec set_vertices(object(), vertices(), usage_hint()) -> {ok, object()} | out_of_memory.

Set the vertices of the mesh.

It sets the vertices.

It's equivalent to set_vertices(Mesh, Vertices, UsageHint, no_copy).

See set_vertices/3 for more details.

set_vertices/4

-spec set_vertices(object(), vertices(), usage_hint(), keep_copy()) -> {ok, object()} | out_of_memory.

Set the vertices of the mesh.

It sets the vertices foobar.

update_vertex(Mesh, Vertex, Offset)

-spec update_vertex(object(), vertex(), slice:offset()) -> ok.

To be written.

To be written.

update_vertices(Mesh, Vertices)

-spec update_vertices(object(), vertices()) -> {ok, object()}.

To be written.

To be written.

update_vertices/3

-spec update_vertices(object(), vertices(), slice:offset()) -> {ok, object()} | out_of_range.

To be written.

To be written.

usage_hint/1

-spec usage_hint(object()) -> usage_hint().

The usage hint of the mesh.

To be written. as set by set_vertices/2.

vertex_count/1

-spec vertex_count(object()) -> non_neg_integer().

The number of vertices in the mesh.

It returns the number of vertices that are currently set in the mesh, as set by set_vertices/2.

with_vertices(Vertices)

-spec with_vertices(vertices()) -> {ok, object()}.

To be written.

To be written.

with_vertices(Vertices, UsageHint)

-spec with_vertices(vertices(), usage_hint()) -> {ok, object()}.

To be written.

To be written.

with_vertices(Vertices, UsageHint, KeepCopy)

-spec with_vertices(vertices(), usage_hint(), keep_copy()) -> {ok, object()}.

To be written.

To be written.