texture (beam_graphics v0.1.0)

Texture

A texture is a 2D array of pixels (or in other words, an image) that can be used for rendering. They are indirectly rendered on a surface with a 2D or 3D mesh that specifies texture coordinates for each vertex.

For instance, the following will render an image.

Mesh = mesh:with_vertices([
    {100.0, 100.0, 1.0, 0.0},
    {700.0, 100.00, 1.0, 1.0},
    {700.0, 500.00, 0.0, 1.0},
    {700.0, 500.00, 0.0, 0.0}
]).
Texture = texture:with_image({800, 600, load_image("image.png")}).
surface:render(Surface, Mesh, triangle_strip, Texture).

Texture pixels must live in the GPU memory in order to be used for rendering, and therefore, reading and updating pixels have an associated cost. Avoid those operations as much as possible, and if frequent reading is required, pixels can be cached locally with the keep_copy option. It explains the local/remote semantics in the API. For instance, local_image/1 will read the pixels from a local copy (if it exists), while remote_image/1 will read the pixels from the GPU memory (which is more expensive).

To create a texture, use the with_image/1 and with_image/2 functions. The texture is later disposed with the destroy/1 function, which will remove the texture from the GPU memory and free the associated resources.

Using a destroyed texture have undefined behavior.

Note that there cannot be "empty" texture and it must be at least 1x1 pixel in size.

Note that it can return 'out of memory' but in practice it hardly happens so you can ignore it. xxx

For convenience, the with_color/3 and with_color/4 functions can be used to create a texture filled with a single color.

To change the image of a texture, which can potentially resize the texture, use the set_image/2 function.

NOTE: The advantage of using set_image/2 over re-creating another texture is that it will keep the internal OpenGL object ID. Use update_image/2 to update the pixels of a texture without changing its size.

Use the with/1 and height/2 functions to retrieve the width and height of the texture.

To retrieve all the pixels of a texture, or a sub-set of them, use the local_image/1, local_image/3, remote_image/1, and remote_image/3 functions. The "local" version of those functions will read the pixels from a locally cached pixels, if a copy was requested at creation time with the keep_copy option. The "remote" version will read the pixels from the GPU memory, which is more expensive.

The local_pixel/3 and remote_pixel/3 functions allows to retrieve a single pixel at a given position in the texture.

NOTE: Those functions implement the slice API which allows you to easily retrieve sub-regions of the texture. For instance, you can you retrieve the the bottom-right pixel of a texture with remote_pixel(Texture, -1, -1).

To

OpenGL Internals

It wraps a texture OpenGL object. To be written.

Use the handle/1 function to retrieve the OpenGL buffer ID. It's created by the root context, which you can retrieve with the graphics_context:inner_context/0 function.

XXX: No possible optimization (like retrieving a sub-region of the texture) unless using OpenGL 4.5 (glGetTextureSubImage).

XXX: Implement copy operations ? XXX: Investigate any other potential OpenGL texture "properties" I'm not ware

 of (swizzle ? texture compare mode? border color? anisotropic filtering?)., etc
multisample textures (anti-aliasing), etc.

XXX: Or just add documentation to show how to run the raw opengl contexts to

switch thowse advanced properties.

XXX: Perhaps swizzle mask and compare mode could be an interesting property

 to b e exposed.

-export([
    set_border_color/2,  % For border handling
    set_swizzle_mask/2,  % Control RGBA channel swizzling
    set_compare_mode/2,  % For shadow mapping
    set_lod_parameters/5,  % LOD bias, min, max
    set_max_anisotropy/2,  % For anisotropic filtering
    create_from_framebuffer/3  % Create texture from current framebuffer
]).

separete repo with cubemap_texture ?

-export([
    with_array/2,  % Create texture array
    with_cubemap/1,  % Create cubemap texture
    update_cubemap_face/3  % Update specific cubemap face
]).

update wrap mode api to suppor thorizontal and vertical

Types

color_space()

-type color_space() :: linear | s_rgb.

To be written.

height()

-type height() :: non_neg_integer().

image()

-type image() :: {width(), height(), pixels()}.

To be written.

keep_copy()

-type keep_copy() :: no_copy | keep_copy.

To be written.

magnification_filter()

-type magnification_filter() :: nearest | linear.

To be written.

minification_filter()

-type minification_filter() ::
          nearest | linear | nearest_mipmap_nearest | linear_mipmap_nearest | nearest_mipmap_linear |
          linear_mipmap_linear.

To be written.

object()

-opaque object()

To be written.

To be written.

pixels()

-type pixels() :: [graphics:color()].

width()

-type width() :: non_neg_integer().

wrap_mode()

-type wrap_mode() :: repeat | clamp_to_edge | mirrored_repeat.

To be written.

Functions

color_space/1

-spec color_space(object()) -> color_space().

To be written.

To be written.

destroy/1

-spec destroy(object()) -> ok.

To be written.

To be written.

generate_mipmap/1

-spec generate_mipmap(object()) -> ok.

To be written.

To be written.

gl_object/1

-spec gl_object(object()) -> gl:texture().

To be written.

To be written.

has_local_copy/1

-spec has_local_copy(object()) -> boolean().

To be written.

To be written.

keep_local_copy/1

-spec keep_local_copy(object()) -> {ok, object()} | already_local_copy.

To be written.

To be written.

local_image/1

-spec local_image(object()) -> undefined | image().

To be written.

To be written.

local_image/3

-spec local_image(object(), slice:range(), slice:range()) ->
                     undefined | no_range | {slice:offset(), slice:offset(), image()}.

To be written.

To be written.

local_pixel/3

-spec local_pixel(object(), slice:index(), slice:index()) ->
                     undefined | out_of_range | {slice:offset(), slice:offset(), graphics:color()}.

To be written.

To be written.

magnification_filter/1

-spec magnification_filter(object()) -> magnification_filter().

To be written.

To be written.

minification_filter/1

-spec minification_filter(object()) -> minification_filter().

To be written.

To be written.

release_local_copy/1

-spec release_local_copy(object()) -> {ok, object()} | no_local_copy.

To be written.

To be written.

remote_image/1

-spec remote_image(object()) -> image().

To be written.

To be written.

remote_image/3

-spec remote_image(object(), slice:range(), slice:range()) ->
                      no_range | {slice:offset(), slice:offset(), image()}.

To be written.

To be written.

remote_pixel/3

-spec remote_pixel(object(), slice:index(), slice:index()) ->
                      out_of_range | {slice:offset(), slice:offset(), graphics:color()}.

To be written.

To be written.

resize(Texture, Size)

-spec resize(object(), {width(), height()}) -> {ok, object()} | out_of_memory.

To be written.

To be written.

resize(Texture, Size, Color)

-spec resize(object(), {width(), height()}, graphics:color()) -> {ok, object()} | out_of_memory.

To be written.

To be written.

resize/4

-spec resize(object(), {width(), height()}, graphics:color(), color_space()) ->
                {ok, object()} | out_of_memory.

To be written.

To be written.

set_image(Texture, Image)

-spec set_image(object(), image()) -> {ok, object()} | out_of_memory.

To be written.

To be written.

set_image/3

-spec set_image(object(), image(), color_space()) -> {ok, object()} | out_of_memory.

To be written.

To be written.

set_magnification_filter/2

-spec set_magnification_filter(object(), magnification_filter()) -> {ok, object()}.

To be written.

To be written.

set_minification_filter/2

-spec set_minification_filter(object(), minification_filter()) -> {ok, object()}.

To be written.

To be written.

set_wrap_mode/3

-spec set_wrap_mode(object(), Axis :: horizontal | vertical, wrap_mode()) -> {ok, object()}.

To be written.

To be written.

size/1

-spec size(object()) -> {width(), height()}.

To be written.

To be written.

update_from_color(Texture, Color)

-spec update_from_color(object(), graphics:color()) -> {ok, object()}.

To be written.

To be written.

update_from_color/3

-spec update_from_color(object(), graphics:color(), {slice:offset(), slice:offset()}) -> {ok, object()}.

To be written.

To be written.

update_from_surface(Texture, Surface)

-spec update_from_surface(object(), graphics:color()) -> {ok, object()}.

To be written.

To be written.

update_from_surface/3

-spec update_from_surface(object(), graphics:color(), {slice:offset(), slice:offset()}) ->
                             {ok, object()}.

To be written.

To be written.

update_from_texture(Texture1, Texture2)

-spec update_from_texture(object(), graphics:texture()) -> {ok, object()}.

To be written.

To be written.

update_from_texture/3

-spec update_from_texture(object(), graphics:texture(), {slice:offset(), slice:offset()}) ->
                             {ok, object()}.

To be written.

To be written.

update_image(Texture, Image)

-spec update_image(object(), image()) -> {ok, object()}.

To be written.

To be written.

update_image/4

-spec update_image(object(), image(), slice:offset(), slice:offset()) -> {ok, object()}.

To be written.

To be written.

update_pixel/4

-spec update_pixel(object(), graphics:color(), slice:offset(), slice:offset()) -> {ok, object()}.

To be written.

To be written.

with_color(Color, Size)

-spec with_color(graphics:color(), {width(), height()}) -> {ok, graphics:texture()}.

To be written.

To be written.

with_color/3

-spec with_color(graphics:color(), {width(), height()}, color_space()) -> {ok, graphics:texture()}.

To be written.

To be written.

with_color/4

-spec with_color(graphics:color(), {width(), height()}, color_space(), keep_copy()) ->
                    {ok, graphics:texture()}.

To be written.

To be written.

with_image(Image)

-spec with_image(image()) -> {ok, graphics:texture()}.

To be written.

To be written.

with_image(Image, ColorSpace)

-spec with_image(image(), color_space()) -> {ok, graphics:texture()}.

To be written.

To be written.

with_image/3

-spec with_image(image(), color_space(), keep_copy()) -> {ok, graphics:texture()}.

To be written.

To be written.

wrap_mode/2

-spec wrap_mode(object(), Axis :: horizontal | vertical) -> wrap_mode().

To be written.

To be written.