vector2 (beam_graphics v0.1.0)

2D Vector

A 2D vector is a pair of numbers that is typically used to represent 2D positions and directions in the Euclidean plane. Because it can also be seen as a 3D vector and a 3x1 matrix if we assume an invisible third component set to 0.0 (also called the homogeneous coordinate), it defines operations with 3x3 matrices.

The data structure of a 2D vector simply is a tuple of 2 floats where the first component is called X and the second component is called Y. Therefore, 2D vectors can be naturally created with the tuple syntax.

V = {1.0, 2.0}.

To access the X components of a 2D vector, use the x/1 function, and to access the Y component, use the y/1 function.

1.0 = vector2:x(V).
2.0 = vector2:y(V).

A 2D vector where X and Y are set to zero is called a zero vector, which can be conveniently created with the zero/0 function.

{0.0, 0.0} = vector2:zero().

A macro is also defined to represent the zero 2D vector. (The graphics.hrl header must be included.)

{0.0, 0.0} = ?VECTOR2_ZERO

The geometrical operations with 2D vectors are implemented. You can normalize them and compute their length (also called magnitude). You can also compute the dot product and cross product with other 2D vectors.

5.0 = vector2:length({3.0, 4.0}).
{0.6, 0.8} = vector2:normalize({3.0, 4.0}).
11.0 = vector2:dot_product({1.0, 2.0}, {3.0, 4.0}).
-2.0 = vector2:cross_product({1.0, 2.0}, {3.0, 4.0}).

The regular mathematical operations are implemented too. You can add and subtract with other 2D vectors as well as multiply with scalars. Note that if you need to multiply with a 3x3 matrix, see the matrix3:dot_vector/2 function.

{4.0, 6.0} = vector2:add({1.0, 2.0}, {3.0, 4.0}).
{-1.0, 1.0} = vector2:subtract({1.0, 2.0}, {2.0, 1.0}).
{2.0, 4.0} = vector2:multiply({1.0, 2.0}, 2.0).

Finally, the 2D vector can be augmented to a 3D vector with the to_vector3/1 function.

Beware that a well-formed 2D vector always contains floats, not integers.

XXX: Consider implementing the 'rotation' operation. XXX: There should be helpers to ensure integers are converted to floats when the user input may contain integers.

Functions

abs/1

-spec abs(graphics:vector2()) -> graphics:vector2().

To be written.

add/2

-spec add(graphics:vector2(), graphics:vector2()) -> graphics:vector2().

Add a 2D vector to another 2D vector.

It adds the second 2D vector to the first 2D vector. Because the operation is commutative, the order has no importance. If the first vector is denoted V1 and the second is V2, it does V1 + V2.

Note that this operation is also called translation.

angle(V1, V2)

-spec angle(graphics:vector3(), graphics:vector3()) -> graphics:angle().

To be written.

ceil/1

-spec ceil(graphics:vector2()) -> graphics:vector2().

To be written.

clamp_length(Vector, Min, Max)

-spec clamp_length(graphics:vector2(), float(), float()) -> graphics:vector2().

To be written.

cross_product/2

-spec cross_product(graphics:vector2(), graphics:vector2()) -> float().

Compute the cross product of two 2D vectors.

It computes the cross product of two 2D vectors. The cross product is a vector that is perpendicular to the plane defined by the two vectors. The cross product is only defined in 3D space, but it can be computed in 2D space by augmenting the 2D vectors to 3D vectors with a Z component set to 0.0.

direction(V1, V2)

-spec direction(graphics:vector2(), graphics:vector2()) -> graphics:vector2().

To be written.

distance(V1, V2)

-spec distance(graphics:vector2(), graphics:vector2()) -> float().

To be written.

divide/2

-spec divide(graphics:vector2(), float()) -> graphics:vector2().

To be written.

dot_product/2

-spec dot_product(graphics:vector2(), graphics:vector2()) -> float().

Compute the dot product of two 2D vectors.

It computes the dot product of two 2D vectors. The dot product is a scalar value that is the result of the sum of the products of the corresponding components of the two vectors.

Note that this operation is also called the inner product.

floor/1

-spec floor(graphics:vector2()) -> graphics:vector2().

To be written.

from_angle(Angle)

-spec from_angle(graphics:angle()) -> graphics:vector2().

To be written.

is_equal_to(V1, V2)

-spec is_equal_to(graphics:vector2(), graphics:vector2()) -> boolean().

To be written.

is_equal_to(V1, V2, Epsilon)

-spec is_equal_to(graphics:vector2(), graphics:vector2(), float()) -> boolean().

To be written.

is_unit(Vector)

-spec is_unit(graphics:vector2()) -> boolean().

To be written.

is_zero/1

-spec is_zero(graphics:vector2()) -> boolean().

To be written.

length(Vector)

-spec length(graphics:vector2()) -> float().

Compute the length of a 2D vector.

It computes the length of a 2D vector. The length of a vector is the distance between its origin and its end point.

Note that it's also called the magnitude of the vector.

lerp/3

-spec lerp(graphics:vector2(), graphics:vector2(), float()) -> graphics:vector2().

To be written.

max/2

-spec max(graphics:vector2(), graphics:vector2()) -> graphics:vector2().

Compute the maximum of two 2D vectors.

It computes the maximum of two 2D vectors. The maximum of two vectors is the vector where each component is the maximum of the corresponding components of the two vectors.

min/2

-spec min(graphics:vector2(), graphics:vector2()) -> graphics:vector2().

Compute the minimum of two 2D vectors.

It computes the minimum of two 2D vectors. The minimum of two vectors is the vector where each component is the minimum of the corresponding components of the two vectors.

multiply/2

-spec multiply(graphics:vector2(), float()) -> graphics:vector2().

Multiply a 2D vector with a scalar.

It multiplies a 2D vector with a scalar. If the vector is noted V and the scalar V, it does S . V. Because this operation is commutative, it's also equivalent to V . S.

Note that this operation is also called scaling.

negate/1

-spec negate(graphics:vector2()) -> graphics:vector2().

To be written.

normalize/1

-spec normalize(graphics:vector2()) -> graphics:vector2().

Normalize a 2D vector.

It normalizes a 2D vector. The length of the vector is computed and the vector is divided by this length. The result is a vector with the same direction but with a length of 1.0.

perpendicular/1

-spec perpendicular(graphics:vector2()) -> graphics:vector2().

To be written.

project(Vector, Onto)

-spec project(graphics:vector2(), graphics:vector2()) -> graphics:vector2().

To be written.

reflect(Vector, Axis)

-spec reflect(graphics:vector2(), graphics:vector2()) -> graphics:vector2().

To be written.

rotate(Vector, Angle)

-spec rotate(graphics:vector2(), graphics:angle()) -> graphics:vector2().

To be written.

round/1

-spec round(graphics:vector2()) -> graphics:vector2().

To be written.

smooth_lerp(Vector1, Vector2, T)

-spec smooth_lerp(graphics:vector2(), graphics:vector2(), float()) -> graphics:vector2().

To be written.

subtract/2

-spec subtract(graphics:vector2(), graphics:vector2()) -> graphics:vector2().

Subtract a 2D vector from another 2D vector.

It subtracts the second 2D vector from the first 2D vector. Because the operation is not commutative, the order has importance. If the first vector is denoted V1 and the second is V2, it does V1 - V2.

Note that this operation is also called translation.

to_angle/1

-spec to_angle(graphics:vector2()) -> graphics:angle().

To be written.

to_vector3/1

-spec to_vector3(graphics:vector2()) -> graphics:vector3().

Augment a 2D vector.

It augments a 2D vector to a 3D vector. The Z component is set to 0.0.

unit(Vector)

-spec unit(graphics:vector2()) -> graphics:vector2().

To be written.

x/1

-spec x(graphics:vector2()) -> float().

The X component of the 2D vector.

It returns the X component of the 2D vector.

y/1

-spec y(graphics:vector2()) -> float().

The Y component of the 2D vector.

It returns the Y component of the 2D vector.

zero()

-spec zero() -> graphics:vector2().

The zero 2D vector.

It constructs a zero 2D vector (both components set to 0.0).

{0.0, 0.0} = vector2:zero().

Note that the ?VECTOR2_ZERO macro can be used instead.