vector3 (beam_graphics v0.1.0)

3D Vector

A 3D vector is a triplet of numbers that is typically used to represent 3D positions and directions in the Euclidean space. Because it can also be seen as a 4D vector and a 4x1 matrix if we assume an invisible fourth component set to 0.0 (also called the homogeneous coordinate), it defines operations with 4x4 matrices.

The data structure of a 3D vector simply is a tuple of 3 floats where the first component is called X, the second component is called Y and the third component is called Z. Therefore, 3D vectors can be naturally created with the tuple syntax.

V = {1.0, 2.0, 3.0}.

To access the X components of a 3D vector, use the x/1 function, to access the Y component, use the y/1 function, and to access the Z component, use the z/1 function.

1.0 = vector3:x(V).
2.0 = vector3:y(V).
3.0 = vector3:z(V).

A 3D vector where X, Y and Z are set to zero is called a zero vector, which can be conveniently created with the zero/0 function.

{0.0, 0.0, 0.0} = vector3:zero().

A macro is also defined to represent the zero 3D vector. (The graphics.hrl header must be included.)

{0, 0} = ?VECTOR3_ZERO

The geometrical operations with 3D vectors are implemented. You can normalize them and compute their length (also called magnitude). You can also compute the dot product and cross product with other 3D vectors.

7.0710678118654755 = vector3:length({3.0, 4.0, 5.0}).
{0.4242640687119285, 0.565685424949238, 0.7071067811865475} = vector3:normalize({3.0, 4.0, 5.0})
32.0 = vector3:dot_product({1.0, 2.0, 3.0}, {4.0, 5.0, 6.0}).
{-3.0, 6.0, -3.0} = vector3:cross_product({1.0, 2.0, 3.0}, {4.0, 5.0, 6.0})

The regular mathematical operations are implemented too. You can add and subtract with other 3D vectors as well as multiply with scalars. Note that if you need to multiply with a 4x4 matrix, see the matrix4:dot_vector/2 function.

{5.0, 7.0, 9.0} = vector3:add({1.0, 2.0, 3.0}, {4.0, 5.0, 6.0}).
{-3.0, -3.0, -3.0} = vector3:subtract({1.0, 2.0, 3.0}, {4.0, 5.0, 6.0}).
{2.0, 4.0, 6.0} = vector3:multiply({1.0, 2.0, 3.0}, 2.0).

Finally, the 3D vector can be reduced to a 2D vector with the to_vector2/1 function.

Beware that a well-formed 3D vector always contains floats, not integers.

XXX: Consider implementing the 'rotation' operation. XXX: There should be helpers to ensure integers are converted to floats when the user input may contain integers.

scale/2 - Explicit scalar multiplication (clearer intent than multiply)
divide/2 - Component-wise division
negate/1 - Returns the negative of the vector

Functions

abs/1

-spec abs(graphics:vector3()) -> graphics:vector3().

To be written.

add/2

-spec add(graphics:vector3(), graphics:vector3()) -> graphics:vector3().

Add a 3D vector to another 3D vector.

It adds the second 3D vector to the first 3D vector. Because the operation is commutative, the order has no importance. If the first vector is denoted V1 and the second is V2, it does V1 + V2.

Note that this operation is also called translation.

angle(V1, V2, Axis)

-spec angle(graphics:vector3(), graphics:vector3(), graphics:vector3()) -> graphics:angle().

To be written.

ceil/1

-spec ceil(graphics:vector3()) -> graphics:vector3().

To be written.

clamp_length(Vector, Min, Max)

-spec clamp_length(graphics:vector3(), float(), float()) -> graphics:vector3().

To be written.

cross_product/2

-spec cross_product(graphics:vector3(), graphics:vector3()) -> graphics:vector2().

Compute the cross product of two 3D vectors.

It computes the cross product of two 3D vectors. The cross product is a vector that is perpendicular to the two vectors. The cross product is computed by multiplying the corresponding components of the two vectors and subtracting the result of the products of the other corresponding components.

direction(V1, V2)

-spec direction(graphics:vector3(), graphics:vector3()) -> graphics:vector3().

To be written.

distance(V1, V2)

-spec distance(graphics:vector3(), graphics:vector3()) -> float().

To be written.

divide/2

-spec divide(graphics:vector3(), float()) -> graphics:vector3().

To be written.

dot_product/2

-spec dot_product(graphics:vector3(), graphics:vector3()) -> float().

Compute the dot product of two 3D vectors.

It computes the dot product of two 3D vectors. The dot product is a scalar value that is the result of the sum of the products of the corresponding components of the two vectors.

Note that this operation is also called the inner product.

floor/1

-spec floor(graphics:vector3()) -> graphics:vector3().

To be written.

from_angle(Angle, Axis)

-spec from_angle(graphics:angle(), graphics:vector3()) -> graphics:vector3().

To be written.

is_equal_to(V1, V2)

-spec is_equal_to(graphics:vector3(), graphics:vector3()) -> boolean().

To be written.

is_equal_to(V1, V2, Epsilon)

-spec is_equal_to(graphics:vector3(), graphics:vector3(), float()) -> boolean().

To be written.

is_unit(Vector)

-spec is_unit(graphics:vector3()) -> boolean().

To be written.

is_zero/1

-spec is_zero(graphics:vector3()) -> boolean().

To be written.

length(Vector)

-spec length(graphics:vector3()) -> float().

Compute the length of a 3D vector.

It computes the length of a 3D vector. The length of a vector is the distance between its origin and its end point.

Note that it's also called the magnitude of the vector.

lerp/3

-spec lerp(graphics:vector3(), graphics:vector3(), float()) -> graphics:vector3().

To be written.

max/2

-spec max(graphics:vector3(), graphics:vector3()) -> graphics:vector3().

Compute the maximum of two 3D vectors.

It computes the maximum of two 3D vectors. The maximum of two vectors is the vector where each component is the maximum of the corresponding components of the two vectors.

min/2

-spec min(graphics:vector3(), graphics:vector3()) -> graphics:vector3().

Compute the minimum of two 3D vectors.

It computes the minimum of two 3D vectors. The minimum of two vectors is the vector where each component is the minimum of the corresponding components of the two vectors.

multiply/2

-spec multiply(graphics:vector3(), float()) -> graphics:vector3().

Multiply a 3D vector with a scalar.

It multiplies a 3D vector with a scalar. If the vector is noted V and the scalar V, it does S . V. Because this operation is commutative, it's also equivalent to V . S.

Note that this operation is also called scaling.

negate/1

-spec negate(graphics:vector3()) -> graphics:vector3().

To be written.

normalize/1

-spec normalize(graphics:vector3()) -> graphics:vector3().

Normalize a 3D vector.

It normalizes a 3D vector. The length of the vector is computed and the vector is divided by this length. The result is a vector with the same direction but with a length of 1.0.

project(Vector, Onto)

-spec project(graphics:vector3(), graphics:vector3()) -> graphics:vector3().

To be written.

reflect(Vector, Axis)

-spec reflect(graphics:vector3(), graphics:vector3()) -> graphics:vector3().

To be written.

rotate(Vector, Angle, Axis)

-spec rotate(graphics:vector3(), graphics:angle(), graphics:vector3()) -> graphics:vector3().

To be written.

round/1

-spec round(graphics:vector3()) -> graphics:vector3().

To be written.

smooth_lerp(Vector1, Vector2, T)

-spec smooth_lerp(graphics:vector3(), graphics:vector3(), float()) -> graphics:vector3().

To be written.

subtract/2

-spec subtract(graphics:vector3(), graphics:vector3()) -> graphics:vector3().

Subtract a 3D vector from another 3D vector.

It subtracts the second 3D vector from the first 3D vector. Because the operation is not commutative, the order has importance. If the first vector is denoted V1 and the second is V2, it does V1 - V2.

Note that this operation is also called translation.

to_angle/2

-spec to_angle(graphics:vector3(), graphics:vector3()) -> graphics:angle().

To be written.

to_vector2/1

-spec to_vector2(graphics:vector3()) -> graphics:vector2().

Reduce a 3D vector.

It reduces a 3D vector to a 2D vector. The Z component is discarded.

unit(Vector)

-spec unit(graphics:vector3()) -> graphics:vector3().

To be written.

x/1

-spec x(graphics:vector3()) -> float().

The X component of the 3D vector.

It returns the X component of the 3D vector.

y/1

-spec y(graphics:vector3()) -> float().

The Y component of the 3D vector.

It returns the Y component of the 3D vector.

z/1

-spec z(graphics:vector3()) -> float().

The Z component of the 3D vector.

It returns the Z component of the 3D vector.

zero()

-spec zero() -> graphics:vector3().

A zero 3D vector.

It constructs a zero 3D vector.